apps/ai/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pillow==11.0.0
python-dotenv==1.0.1
httpx==0.28.1
python-multipart==0.0.20
